Added dlib objects and constructors to enable dlib recognitiong with Opencv detection
Added dlib objects in order to be able to do detection with OpenCV but recognition with Dlib. Dlib is missing a pythong binding to construct the list returned after the detection so I added as globals, some dlib objects even though I'm not using them for detection. Dlib objects are cnn_face_detector, mmod_rectangles and img.