added a working embedding algo from OpenFace project
This small OpenFace embedding is fast enough to work  when I run it on the nano, but only when the confidence is high. The plan is to run recognition only in this instances, but for the remaining of the appereance of the subject I can run tracking with machine learning which should be faster and avoids running recognition on every frame. But I'll do the tracking last.