Added archiconda.sh and instructions to install scikit-learn, a fast training model
In order to add scikit-learn to the nano, after struggling with the lack of a wheel compier for fortran nor with PEP 517, required to build several scipy packages, I ended  up using the version of conda already tested on the nano. I downloade it from the nvidia forum, and added the file to the project.

I've also wrote a training model but is not ready yet. Right now it just parses the folder where the training data will be separated by folder name.